// mydev/MyProj.CM/WebApp/campaignmanager/src/app/shared/enum/em-application-module.enum.ts
export enum ApplicationModule {
    connect_media = 'cmpwebapp',
    audiences = 'audiences',
    nominations = 'nominations',
    events = 'events',
    admin = 'admin',
    dashboard = 'dashboard'
}

export enum ApplicationModuleTabs {
    dashboard = 1,
    audiences = 2,
    nominations = 3,
    events = 4,
    admin = 5
}